chore(main): release 1.1.3-beta

+++ agents/publish/publish-website.mjs
@@ -167,6 +167,15 @@ export default async function publishWebsite(
   await fs.mkdir(pagesDir, {
     recursive: true,
   });
+
+  // check outputDir is exist
+  if (!fs.existsSync(outputDir)) {
+    return {
+      message:
+        "It seems that the pages does not generated. \nPlease generate the pages as needed using \`aigne web generate\`\n",
+    };
+  }
+
   await fs.cp(outputDir, pagesDir, {
     recursive: true,
   });
